Large change and update
CPU is not complete but further along. Extension devices are coming along. You can now get interrupts from the keyboard and read it's register, but it's not correct.  It just sits there and doesn't clear.  You can write better programs to take advantage of this.

Programming language processor is coming along.  You can set flags at the head of your file to determine the type of file to be written.  Working on software to program LSAssembly and convert to cpu code.